GP-38956 Fix rounding inconsistency when calculating runtime
This fixes an inconsistent use of floor rounding for the subtrahend
and minuend when calculating runtimes, potentially causing negative
runtimes to be returned (which cannot be saved in unsigned int columns).

+++ CRM/Sqltasks/Task.php
@@ -364,7 +364,7 @@ class CRM_Sqltasks_Task {
     $this->status = 'running';
     $this->error_count = 0;
     $this->reset();
-    $task_timestamp = microtime(TRUE) * 1000;
+    $task_timestamp = (int) (microtime(TRUE) * 1000);
     $this->log('Start running task!', 'info', TRUE);
 
     if ($this->isArchived()) {
